refactor vertex & gemini
